moved keycloak user jobs
